requirements.txt
cachetools >= 2.0.0
librosa >= 0.7.0
pandas >= 0.18.0
pyannote.core >= 4.1
pyannote.database >= 4.0
pyannote.metrics >= 2.3
pyannote.pipeline >= 1.5
pyYAML >= 3.12
scikit-learn >= 0.20.2
sortedcollections >= 1.0.1
sortedcontainers >= 2.0.4
soundfile >= 0.10.2
tqdm >= 4.29.1
tensorboard >= 2.0.0
typing_extensions >= 3.7.4;python_version < '3.8'
pescador >= 2.1.0
Pillow >= 6.2.1
pytorch-lightning >= 0.8.0
pytorch-metric-learning == 0.9.87